requirements.txt
firebase_admin
streamlit
streamlit_option_menu
scikit-learn
polars
matplotlib
seaborn
#python -m pip install --upgrade pip setuptools wheel
scikit-surprise